Implement logic for Tic-Tac-Toe in console
This commit implements all of the game logic for
Tic-Tac-Toe, and is available to play in the
console. Players can type "game.playRound(row, col)"
into the console, where row and col are the the row
and column of the game board respectively. The game
can accommodate two players, where the outcome of
a match is a win or tie.